index on master: d67a8c0 Add `--locked` v0.7.1 (#10)
rebase
